Stage5_OpenSpec_Applied/Lesson09_OpenSpec_Advanced_Practice/my-website/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
alembic>=1.14.0
pyjwt>=2.10.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
python-dotenv>=1.1.0
